# Optimize Symbol to_msgpack_ext and from_msgpack_ext