@tagomoris 2021-02-16
This is a completely off-topic question: Why is the benchmark on Ruby 3.0 so slower than Ruby 2.7?