@casperisfine 2021-02-16
Hum, it's probably just my ruby 3.0 that's compiled in debug mode:

```ruby
require 'benchmark/ips'
str = :foobar.to_s
Benchmark.ips do |x|
  x.report('to_sym') { str.to_sym }
end
```

```
$ chruby_use /opt/rubies/3.0.0-pshopify3/
$ ruby /tmp/bench-symbols.rb 
Warming up --------------------------------------
              to_sym   202.948k i/100ms
Calculating -------------------------------------
              to_sym      2.011M (± 1.1%) i/s -     10.147M in   5.046170s
$ chruby_use /opt/rubies/2.7.2/
$ ruby /tmp/bench-symbols.rb 
Warming up --------------------------------------
              to_sym     1.268M i/100ms
Calculating -------------------------------------
              to_sym     12.581M (± 0.9%) i/s -     63.419M in   5.041164s
```