@tagomoris 2021-02-16
I'm ok to merge this change, but I'm feeling that the original code is also important to show how it should be (without optimization).
@casperisfine How do you feel if I say I want to leave the original code as comments?