@casperisfine 2021-02-16
>  I'm feeling that the original code is also important to show how it should be 

You mean showing how extensions are supposed to use `Array#pack` and `String.unpack`? I don't think the Symbol serializer is the best place for that. 

This sounds to me that it would be more the job of the README, no? https://github.com/msgpack/msgpack-ruby#extension-types