@tagomoris 2021-02-16
@casperisfine It's not what I meant. My intention was, the code of this change is highly relying on the current MRI's implementation, and it can be mapped to the MessagePack's requirements.
The original code (using pack/unpack) is independent of internal implementation. So, I feel that it may be better to show the original code (as comments) and the newer optimized code in parallel to explain what the optimized code is doing.
Does it make sense?