@casperisfine 2021-02-16
> the code of this change is highly relying on the current MRI's implementation

I don't see how. The `String <-> Symbol` casting with `.to_sym` / `to_s|to_name` is standard, and will work the same on Truffle, JRuby etc.

Since MessagePack extensions basically expose strings, I suppose it's fine to directly use theses. Well except for https://github.com/msgpack/msgpack-ruby/pull/211 but not sure how we could ever fix that one without breaking backward compat.

SO yeah, not sure I understand your point.