@tagomoris 2021-02-16
MessagePack is a binary serialization format, and **I feel** we should use `pack` to get the binary representations of objects, generally speaking.
Of course, I know that we can use string values directly in this case, but also want to show how it can be written using `pack`/`unpack` at the same time.