@casperisfine 2021-02-16
As you wish, I'll restore the old code as comment.