@tagomoris 2021-02-16
@casperisfine Thank you! The added explanation is perfect!
Usually, I strongly hate the behavior to leave older code as comments, but this case was something different to me.