@casperisfine 2021-02-16
Well, it's more like inline documentation in this case so yeah I understand.